Rename test/jsdom/index.js to test/jsdom/misc.js
