+++ .jscsrc
@@ -44,7 +44,7 @@
     "test/browser/index.js",
     "test/jquery-fixtures",
     "test/jsdom/files",
-    "test/jsdom/index.js",
+    "test/jsdom/misc.js",
     "test/jsdom/leak.js",
     "test/level1",
     "test/level2",
